producer/requirements.txt
confluent-kafka==2.3.0
websocket-client==1.6.4
requests==2.31.0
